+++ src/app/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './chat.provider';
 export * from './feedback.provider';
+export * from './list-files.providers';