+++ src/app/providers/list-files.providers.ts
@@ -7,27 +7,26 @@ import {
   TreeItem,
 } from 'vscode';
 
-import { EXTENSION_ID } from '../configs';
-import { FeedbackController } from '../controllers';
+import { ListFilesController } from '../controllers';
 import { NodeModel } from '../models';
 
 /**
- * The FeedbackProvider class
+ * The ListFilesProvider class
  *
  * @class
- * @classdesc The class that represents the feedback provider.
+ * @classdesc The class that represents the list of files provider.
  * @export
  * @public
  * @implements {TreeDataProvider<NodeModel>}
  * @property {EventEmitter<NodeModel | undefined | null | void>} _onDidChangeTreeData - The onDidChangeTreeData event emitter
  * @property {Event<NodeModel | undefined | null | void>} onDidChangeTreeData - The onDidChangeTreeData event
- * @property {FeedbackController} controller - The feedback controller
+ * @property {ListFilesController} controller - The list of files controller
  * @example
- * const provider = new FeedbackProvider();
+ * const provider = new ListFilesProvider();
  *
  * @see https://code.visualstudio.com/api/references/vscode-api#TreeDataProvider
  */
-export class FeedbackProvider implements TreeDataProvider<NodeModel> {
+export class ListFilesProvider implements TreeDataProvider<NodeModel> {
   // -----------------------------------------------------------------
   // Properties
   // -----------------------------------------------------------------
@@ -37,7 +36,7 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
    * The onDidChangeTreeData event emitter.
    * @type {EventEmitter<NodeModel | undefined | null | void>}
    * @private
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    * @example
    * this._onDidChangeTreeData = new EventEmitter<Node | undefined | null | void>();
    * this.onDidChangeTreeData = this._onDidChangeTreeData.event;
@@ -53,7 +52,7 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
    * The onDidChangeTreeData event.
    * @type {Event<NodeModel | undefined | null | void>}
    * @public
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    * @example
    * readonly onDidChangeTreeData: Event<Node | undefined | null | void>;
    * this.onDidChangeTreeData = this._onDidChangeTreeData.event;
@@ -67,13 +66,13 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
   // -----------------------------------------------------------------
 
   /**
-   * Constructor for the FeedbackProvider class
+   * Constructor for the ListFilesProvider class
    *
    * @constructor
    * @public
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    */
-  constructor(readonly controller: FeedbackController) {
+  constructor(readonly controller: ListFilesController) {
     this._onDidChangeTreeData = new EventEmitter<
       NodeModel | undefined | null | void
     >();
@@ -91,7 +90,7 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
    * @function getTreeItem
    * @param {NodeModel} element - The element
    * @public
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    * @example
    * const treeItem = provider.getTreeItem(element);
    *
@@ -109,7 +108,7 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
    * @function getChildren
    * @param {NodeModel} [element] - The element
    * @public
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    * @example
    * const children = provider.getChildren(element);
    *
@@ -122,7 +121,7 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
       return element.children;
     }
 
-    return this.getFeedbacks();
+    return this.getListFiles();
   }
 
   /**
@@ -142,42 +141,42 @@ export class FeedbackProvider implements TreeDataProvider<NodeModel> {
 
   // Private methods
   /**
-   * Returns the feedbacks.
+   * Gets the list of files.
    *
-   * @function getFeedbacks
+   * @function getListFiles
    * @private
-   * @memberof FeedbackProvider
+   * @memberof ListFilesProvider
    * @example
-   * const feedbacks = this.getFeedbacks();
+   * const files = provider.getListFiles();
    *
-   * @returns {NodeModel[]} - The feedbacks
+   * @returns {Promise<NodeModel[]>} - The list of files
    */
-  private getFeedbacks(): NodeModel[] {
-    return [
-      new NodeModel('About Us', new ThemeIcon('info'), {
-        title: 'About Us',
-        command: `${EXTENSION_ID}.feedback.aboutUs`,
-      }),
-      new NodeModel('Documentation', new ThemeIcon('book'), {
-        title: 'Documentation',
-        command: `${EXTENSION_ID}.feedback.documentation`,
-      }),
-      new NodeModel('Report Issues', new ThemeIcon('bug'), {
-        title: 'Report Issues',
-        command: `${EXTENSION_ID}.feedback.reportIssues`,
-      }),
-      new NodeModel('Rate Us', new ThemeIcon('star'), {
-        title: 'Rate Us',
-        command: `${EXTENSION_ID}.feedback.rateUs`,
-      }),
-      new NodeModel('Follow Us', new ThemeIcon('twitter'), {
-        title: 'Follow Us',
-        command: `${EXTENSION_ID}.feedback.followUs`,
-      }),
-      new NodeModel('Support Us', new ThemeIcon('heart'), {
-        title: 'Support Us',
-        command: `${EXTENSION_ID}.feedback.supportUs`,
-      }),
-    ];
+  private async getListFiles(): Promise<NodeModel[]> {
+    const files = await this.controller.getFiles();
+
+    if (!files) {
+      return [];
+    }
+
+    const nodes: NodeModel[] = [];
+
+    const fileTypes = this.controller.config.include;
+
+    for (const fileType of fileTypes) {
+      const children = files.filter((file) =>
+        file.label.toString().endsWith(`.${fileType}`),
+      );
+      const node = new NodeModel(
+        `${fileType}: ${children.length}`,
+        new ThemeIcon('folder-opened'),
+        undefined,
+        undefined,
+        fileType,
+        children,
+      );
+      nodes.push(node);
+    }
+
+    return nodes;
   }
 }